Removed references to outdated service URLs
Public API gateway in front of this lambda function has been
replaced with a private API gateway. Domain name and address of
the private API gateway does not need to be kept in this repository
anymore - the API is only accessed from security servers
(liityntäpalvelin) and the address is configured manually in
security server's UI. When WSDL is retrived by "palveluväylä",
the address is replaced with another example address before
publishing. Thus, it does not matter what address value is
specified here.